requirements.txt
%%writefile requirements.txt
streamlit
pyngrok
pandas
torch
sentence-transformers
pdfplumber
python-docx
unidecode
